pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.2",
    "celery>=5.5.3",
    "fastapi>=0.115.13",
    "opencv-python>=4.11.0.86",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pytesseract>=0.3.13",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.3",
]
